Refactor hooks into components; create new empty files for future features
