+++ src/components/Waveform.tsx
@@ -1,0 +1,2 @@
+// SMALL TIME BUFFER (think 2-20 ms, avoids looking like its a scrolling scope) (X axis)
+// Amplitude (Y axis)
