+++ src/components/XY.tsx
@@ -1,0 +1,5 @@
+// 'lissajous scope', 'vectorscope', 'phase scope', 'goniometer'
+// one axis: left channel
+// other axis: right channel
+// mono would appear vertical
+// full cancelation would appear horizontal???
